完成users组件所有功能

+++ src/components/users/users.vue
@@ -10,41 +10,42 @@
       <!-- 搜索框 -->
       <div style="margin: 20px 0">
         <el-input placeholder="请输入内容" class="searchUser" v-model="query" clearable>
+          <!-- 搜索按钮 -->
           <el-button @click="searchUsers()" slot="append" icon="el-icon-search"></el-button>
         </el-input>
-        <!--  添加用户 按钮 -->
-        <el-button class="sureUserAdd" @click="addDialogFormVisible = true" type="success" plain>添加用户</el-button>
+        <!-- 添加用户 按钮 -->
+        <el-button class="sureUserAdd" @click="addUser()" type="success" plain>添加用户</el-button>
       </div>
-      <!-- 表格主体 -->
-      <el-table :data='userList' stripe style="width: 100%">
-        <el-table-column prop="id" label="#" width="50">
-        </el-table-column>
+      <!-- 用户列表  表格主体 -->
+      <el-table :data="userList" stripe style="width: 100%">
+        <el-table-column prop="id" label="#" width="50"> </el-table-column>
         <el-table-column prop="username" label="姓名" width="110">
         </el-table-column>
-        <el-table-column prop="email" label="邮箱">
-        </el-table-column>
+        <el-table-column prop="email" label="邮箱"> </el-table-column>
         <el-table-column prop="mobile" label="电话" width="180">
         </el-table-column>
         <el-table-column prop="create_time" label="创建日期" width="180">
           <template slot-scope="scope">
-            {{scope.row.create_time | fmtdate}}
+            {{ scope.row.create_time | fmtdate }}
           </template>
         </el-table-column>
-
+        <!-- 改变用户状态 -->
         <el-table-column label="用户状态" width="180">
           <template slot-scope="scope">
-            <el-switch v-model="scope.row.mg_state" active-color="#13ce66" inactive-color="#ff4949">
+            <el-switch @change="stateChange(scope.row)" v-model="scope.row.mg_state" active-color="#13ce66"
+              inactive-color="#ff4949">
             </el-switch>
           </template>
         </el-table-column>
 
         <el-table-column label="操作" width="180">
           <template slot-scope="scope">
-
-            <el-button @click="editUser(scope.row.id)" type="primary" size="mini" icon="el-icon-edit" circle plain></el-button>
+            <!-- 编辑按钮 -->
+            <el-button @click="editUser(scope.row)" type="primary" size="mini" icon="el-icon-edit" circle plain></el-button>
+            <!-- 删除按钮 -->
             <el-button @click="deleteUser(scope.row.id)" type="danger" size="mini" icon="el-icon-delete" circle plain></el-button>
-            <el-button type="success" size="mini" icon="el-icon-check" circle plain></el-button>
-
+            <!-- 分配角色按钮 -->
+            <el-button @click="roleChange(scope.row)" type="success" size="mini" icon="el-icon-check" circle plain></el-button>
           </template>
         </el-table-column>
       </el-table>
@@ -79,7 +80,7 @@
       <el-dialog title="编辑用户" :visible.sync="editDialogFormVisible">
         <el-form :model="form">
           <el-form-item label="用户名" :label-width="formLabelWidth">
-            <el-input v-model="form.username" autocomplete="off"></el-input>
+            <el-input v-model="form.username" autocomplete="off" :disabled="true"></el-input>
           </el-form-item>
           <el-form-item label="邮箱" :label-width="formLabelWidth">
             <el-input v-model="form.email" autocomplete="off"></el-input>
@@ -93,6 +94,26 @@
           <el-button type="primary" @click="sureUserEdit()">确 定</el-button>
         </div>
       </el-dialog>
+      <!-- "分配角色"对话框 -->
+      <el-dialog title="分配角色" :visible.sync="roleDialogFormVisible">
+        <el-form :model="form">
+          <el-form-item label="用户名" label-width="100px">
+            {{ currUsername }}
+          </el-form-item>
+          <el-form-item label="角色" label-width="100px">
+            <!-- select绑定的值和option中的value相同时显示相应文本 -->
+            <el-select v-model="currRoleId">
+              <!-- 遍历roles数组 -->
+              <el-option :value="-1" label="请选择角色" disabled="true"></el-option>
+              <el-option v-for="(item,index) in roles" :key="index" :label="item.roleName" :value="item.id"></el-option>
+            </el-select>
+          </el-form-item>
+        </el-form>
+        <div slot="footer" class="dialog-footer">
+          <el-button @click="roleDialogFormVisible = false">取 消</el-button>
+          <el-button type="primary" @click="sureChangeRole()">确 定</el-button>
+        </div>
+      </el-dialog>
     </el-card>
   </div>
 </template>
@@ -101,27 +122,32 @@
     data() {
       return {
         //查询参数
-        query: '',
+        query: "",
         pagenum: 1,
         pagesize: 2,
         total: 0,
         userList: [{
-          username: '',
-          email: '',
-          mobile: '',
-          create_time: ''
+          username: "",
+          email: "",
+          mobile: "",
+          create_time: ""
         }],
         isShowClear: false,
         form: {
-          username: '',
-          password: '',
-          email: '',
-          mobile: ''
+          username: "",
+          password: "",
+          email: "",
+          mobile: ""
         },
         addDialogFormVisible: false,
         editDialogFormVisible: false,
-        formLabelWidth: '100px',
-
+        roleDialogFormVisible: false,
+        formLabelWidth: "100px",
+        // 分配角色
+        currUsername: "",
+        currRoleId: -1,
+        roles: [],
+        currUserId: ''
       }
     },
 
@@ -129,15 +155,61 @@
       this.getUsers()
     },
     methods: {
+
+      // 选择/更改角色 打开对话框
+      async roleChange(user) {
+        this.roleDialogFormVisible = true
+
+        this.currUsername = user.username
+        this.currUserId = user.id
+        // 发送请求 获取全部的roles 
+        var res1 = await this.$http.get('roles')
+        // console.log(res1)
+        // data:
+        //   data: Array(5)
+        //        0:
+        //          children: (...)
+        //          id: 30
+        //          roleDesc: "技术负责人"
+        //          roleName: "主管"
+        // 把数组赋值给roles 主要是需要里面的id(角色id) (显示在option中的:value)
+        this.roles = res1.data.data
+        var res2 = await this.$http.get(`users/${user.id}`)
+        // console.log(res2)
+        this.currRoleId = res2.data.data.rid
+        // data:
+        //  data:
+        //    email: "123@qq.com"
+        //    id: 502
+        //    mobile: "1111111111111111"
+        //    rid: 34
+        //    username: "linken"
+        // 需要里面的rid  用作当前用户 的角色id -> currRoleId
+      },
+      // 确认更改/选择角色
+      async sureChangeRole() {
+        this.roleDialogFormVisible = false
+        var res = this.$http.put(`users/${this.currUserId}/role`, {
+          rid: this.currRoleId
+        })
+        console.log(res)
+      },
+      // 改变用户状态
+      // users/:uId/state/:type
+      async stateChange(user1) {
+        this.$http.put(`users/${user1.id}/state/${user1.mg_state}`)
+        // console.log(res)
+      },
       // 编辑用户
-      editUser() {
+      editUser(user) {
+        this.form = user
         this.editDialogFormVisible = true
       },
       // 确认用户编辑完成
-      async sureUserEdit(edituserID) {
-        this.editDialogFormVisible = false
-        const res = await this.$http.put(`users/${edituserID}`)
-
+      async sureUserEdit() {
+        this.editDialogFormVisible = false;
+        const res = await this.$http.put(`users/${this.form.id}`, this.form)
+        console.log(this.form)
         const {
           meta: {
             msg,
@@ -149,8 +221,6 @@
         if (status === 200) {
           this.$message.success(msg)
         }
-
-
       },
       // 搜索按钮
       searchUsers() {
@@ -160,31 +230,43 @@
         this.getUsers()
       },
       // 分页相关方法
-      handleSizeChange(val) { /* val哪来的 */
-        console.log(`每页${val}条数据`);
+      handleSizeChange(val) {
+        // val来源于被点击的标签 pagesize
+        console.log(val)
+        console.log(`每页${val}条数据`)
         this.pagesize = val
         // 重置页码
         this.pagenum = 1
         this.getUsers()
+        console.log(this.pagenum)
       },
       handleCurrentChange(val) {
-        console.log(`当前是第${val}页`);
+        // val来源于被点击的标签 pagenum
+        console.log(val)
+        console.log(`当前是第${val}页`)
         this.pagenum = val
         this.getUsers()
+
+
+      },
+      // 添加用户
+      addUser() {
+        this.addDialogFormVisible = true
+        this.form = {}
       },
       // 确认添加用户
       async sureUserAdd() {
         // 关闭窗口
         this.addDialogFormVisible = false
         // 发送异步请求
-        var res = await this.$http.post('users', this.form)
+        var res = await this.$http.post("users", this.form)
         const {
           meta: {
             status,
             msg
           },
           data
-        } = res.data
+        } = res.data;
         if (status === 201) {
           console.log(res)
           this.$message.success(msg)
@@ -198,43 +280,48 @@
       },
       // 1111111111111111111111111111111111111111111111111111
       deleteUser(userID) {
-        this.$confirm('此操作将永久删除该文件, 是否继续?', '提示', {
-          confirmButtonText: '确定',
-          cancelButtonText: '取消',
-          type: 'warning'
-        }).then(async () => {
-          // 发送删除请求
-          const res = await this.$http.delete(`users/${userID}`)
-          const {
-            meta: {
-              msg,
-              status
-            },
-            data
-          } = res.data
-          console.log(res)
-          if (status === 200) {
+        this.$confirm("此操作将永久删除该文件, 是否继续?", "提示", {
+            confirmButtonText: "确定",
+            cancelButtonText: "取消",
+            type: "warning"
+          })
+          .then(async () => {
+            // 发送删除请求
+            const res = await this.$http.delete(`users/${userID}`)
+            const {
+              meta: {
+                msg,
+                status
+              },
+              data
+            } = res.data
+
+            if (status === 200) {
+              this.$message({
+                type: "success",
+                message: msg
+              })
+              this.pagenum = 1
+              this.getUsers()
+            }
+          })
+          .catch(() => {
             this.$message({
-              type: 'success',
+              type: "info",
               message: msg
             })
-            this.getUsers()
-          }
-        }).catch(() => {
-          this.$message({
-            type: 'info',
-            message: msg
           })
-        })
-
       },
       // 获取用户列表
       async getUsers() {
         // 接口文档要求授权API->设置请求头
-        const AUTH_TOKEN = localStorage.getItem('token')
-        this.$http.defaults.headers.common['Authorization'] = AUTH_TOKEN
+        const AUTH_TOKEN = localStorage.getItem("token")
+        this.$http.defaults.headers.common["Authorization"] = AUTH_TOKEN
         const res = await this.$http.get(
-          `users?query=${this.query}&pagenum=${this.pagenum}&pagesize=${this.pagesize}`)
+          `users?query=${this.query}&pagenum=${this.pagenum}&pagesize=${
+            this.pagesize
+          }`
+        );
         // console.log(res);
         const {
           meta: {
@@ -250,13 +337,13 @@
         if (status === 200) {
           this.total = total
           this.userList = users
-          this.$message.success(msg);
+          this.$message.success(msg)
         } else {
-          this.$message.warning(msg);
+          this.$message.warning(msg)
         }
       }
     }
-  }
+  };
 
 </script>
 <style>
